# static/swagger.json
{
	"swagger": "2.0",
	"info": {
	  "title": "Library Management API",
	  "version": "1.0.0",
	  "description": "API documentation for the Library Management System."
	},
	"host": "127.0.0.1:5000",
	"basePath": "/",
	"schemes": ["http"],
	"paths": {
	  "/": {
		"get": {
		  "summary": "Home Page",
		  "description": "Show the main page of the application.",
		  "responses": {
			"200": {
			  "description": "Successful operation"
			}
		  }
		}
	  },
	  "/books": {
		"post": {
		  "summary": "Add a New Book",
		  "description": "Adds a new book to the library.",
		  "consumes": ["application/json"],
		  "parameters": [
			{
			  "in": "body",
			  "name": "body",
			  "description": "Book object that needs to be added to the library",
			  "required": true,
			  "schema": {
				"$ref": "#/definitions/Book"
			  }
			}
		  ],
		  "responses": {
			"201": {
			  "description": "Book added successfully"
			},
			"400": {
			  "description": "Missing required fields"
			}
		  }
		},
		"get": {
		  "summary": "List All Books",
		  "description": "Retrieves a list of all books currently in the library.",
		  "produces": ["application/json"],
		  "responses": {
			"200": {
			  "description": "A list of books.",
			  "schema": {
				"type": "array",
				"items": {
				  "$ref": "#/definitions/Book"
				}
			  }
			}
		  }
		}
	  },
	  "/books/search": {
		"get": {
		  "summary": "Search Books",
		  "description": "Searches for books based on a query. You may filter by author, published year, and genre.",
		  "produces": ["application/json"],
		  "parameters": [
			{
			  "name": "author",
			  "in": "query",
			  "description": "Author of the book",
			  "required": false,
			  "type": "string"
			},
			{
			  "name": "published_year",
			  "in": "query",
			  "description": "Published year of the book",
			  "required": false,
			  "type": "integer",
			  "format": "int32"
			},
			{
			  "name": "genre",
			  "in": "query",
			  "description": "Genre of the book",
			  "required": false,
			  "type": "string"
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Search results",
			  "schema": {
				"type": "array",
				"items": {
				  "$ref": "#/definitions/Book"
				}
			  }
			}
		  }
		}
	  },
	  "/books/{isbn}": {
		"delete": {
		  "summary": "Delete a Book",
		  "description": "Deletes a book from the library based on its ISBN.",
		  "parameters": [
			{
			  "name": "isbn",
			  "in": "path",
			  "description": "ISBN of the book to delete",
			  "required": true,
			  "type": "string"
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Book deleted successfully"
			}
		  }
		},
		"put": {
		  "summary": "Update a Book",
		  "description": "Updates the details of an existing book identified by its ISBN.",
		  "consumes": ["application/json"],
		  "parameters": [
			{
			  "name": "isbn",
			  "in": "path",
			  "description": "ISBN of the book to update",
			  "required": true,
			  "type": "string"
			},
			{
			  "in": "body",
			  "name": "body",
			  "description": "Updated book object",
			  "required": true,
			  "schema": {
				"$ref": "#/definitions/Book"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Book updated successfully."
			},
			"404": {
			  "description": "Book not found."
			}
		  }
		}
	  }
	},
	"definitions": {
	  "Book": {
		"type": "object",
		"required": ["title", "author", "published_year", "isbn"],
		"properties": {
		  "title": {
			"type": "string"
		  },
		  "author": {
			"type": "string"
		  },
		  "published_year": {
			"type": "integer",
			"format": "int32"
		  },
		  "isbn": {
			"type": "string"
		  },
		  "genre": {
			"type": "string"
		  }
		}
	  }
	}
  }